pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "pydantic-ai>=1.0.8",
    "python-dotenv>=1.1.1",
]
